♻️ [RUMF-1267] remove cycle in recorder
Move `shouldIgnoreNode` to `privacy` module. `shouldIgnoreNode` is only
used in `privacy`, so it makes sense to put it there.
