+++ packages/rum/src/domain/record/serialize.ts
@@ -174,79 +174,6 @@ export function serializeElementNode(element: Element, options: SerializeOptions
 }
 
 /**
- * TODO: Preserve CSS element order, and record the presence of the tag, just don't render
- * We don't need this logic on the recorder side.
- * For security related meta's, customer can mask themmanually given they
- * are easy to identify in the HEAD tag.
- */
-export function shouldIgnoreElement(element: Element): boolean {
-  if (element.nodeName === 'SCRIPT') {
-    return true
-  }
-
-  if (element.nodeName === 'LINK') {
-    const relAttribute = getLowerCaseAttribute('rel')
-    return (
-      // Scripts
-      (relAttribute === 'preload' && getLowerCaseAttribute('as') === 'script') ||
-      // Favicons
-      relAttribute === 'shortcut icon' ||
-      relAttribute === 'icon'
-    )
-  }
-
-  if (element.nodeName === 'META') {
-    const nameAttribute = getLowerCaseAttribute('name')
-    const relAttribute = getLowerCaseAttribute('rel')
-    const propertyAttribute = getLowerCaseAttribute('property')
-    return (
-      // Favicons
-      /^msapplication-tile(image|color)$/.test(nameAttribute) ||
-      nameAttribute === 'application-name' ||
-      relAttribute === 'icon' ||
-      relAttribute === 'apple-touch-icon' ||
-      relAttribute === 'shortcut icon' ||
-      // Description
-      nameAttribute === 'keywords' ||
-      nameAttribute === 'description' ||
-      // Social
-      /^(og|twitter|fb):/.test(propertyAttribute) ||
-      /^(og|twitter):/.test(nameAttribute) ||
-      nameAttribute === 'pinterest' ||
-      // Robots
-      nameAttribute === 'robots' ||
-      nameAttribute === 'googlebot' ||
-      nameAttribute === 'bingbot' ||
-      // Http headers. Ex: X-UA-Compatible, Content-Type, Content-Language, cache-control,
-      // X-Translated-By
-      element.hasAttribute('http-equiv') ||
-      // Authorship
-      nameAttribute === 'author' ||
-      nameAttribute === 'generator' ||
-      nameAttribute === 'framework' ||
-      nameAttribute === 'publisher' ||
-      nameAttribute === 'progid' ||
-      /^article:/.test(propertyAttribute) ||
-      /^product:/.test(propertyAttribute) ||
-      // Verification
-      nameAttribute === 'google-site-verification' ||
-      nameAttribute === 'yandex-verification' ||
-      nameAttribute === 'csrf-token' ||
-      nameAttribute === 'p:domain_verify' ||
-      nameAttribute === 'verify-v1' ||
-      nameAttribute === 'verification' ||
-      nameAttribute === 'shopify-checkout-api-token'
-    )
-  }
-
-  function getLowerCaseAttribute(name: string) {
-    return (element.getAttribute(name) || '').toLowerCase()
-  }
-
-  return false
-}
-
-/**
  * Text Nodes are dependant on Element nodes
  * Privacy levels are set on elements so we check the parentElement of a text node
  * for privacy level.
